+++ client/src/commandHandlers.ts
@@ -18,6 +18,12 @@ import * as fs from 'fs';
 import * as path from 'path';
 import { TypeDefinitionRequest } from 'vscode-languageclient';
 
+let outputChannel: vscode.OutputChannel
+
+export function setOutputChannel(oc : vscode.OutputChannel) {
+	outputChannel = oc;
+}
+
 export async function exportArchive(file: vscode.Uri) {
 	try {
 		// HACK - we load the module lazily so that process.browser is set 
@@ -74,6 +80,85 @@ export async function exportClassDiagram(file: vscode.Uri) {
 
 		modelManager.accept(visitor, parameters);
 		vscode.window.showInformationMessage(`Exported class diagram to ${outputPath}`);
+		return true;
+	}
+	catch(error) {
+		vscode.window.showErrorMessage(`Error ${error}`);
+	}
+
+	return false;
+}
+
+export async function triggerClause(file: vscode.Uri) {
+	try {
+		// HACK - we load the module lazily so that process.browser is set 
+		// (see extension.ts)
+		outputChannel.show();
+
+		const Template = require('@accordproject/cicero-core').Template;
+		const Clause = require('@accordproject/cicero-core').Clause;
+		const Engine = require('@accordproject/cicero-engine').Engine;
+
+		const template = await Template.fromDirectory(file.path);
+		const clause = new Clause(template);
+
+		const samplePath = path.join(file.path, 'text', 'sample.md');
+
+		if(!fs.existsSync(samplePath)) {
+			vscode.window.showErrorMessage('Cannot trigger: /text/sample.md file was not found.');
+			return false;
+		}
+
+		const requestPath = path.join(file.path, 'request.json');
+
+		if(!fs.existsSync(requestPath)) {
+			vscode.window.showErrorMessage('Cannot trigger:/request.json file was not found.');
+			return false;
+		}
+
+		const statePath = path.join(file.path, 'state.json');
+
+		if(!fs.existsSync(statePath)) {
+			vscode.window.showErrorMessage('Cannot trigger: /state.json file was not found.');
+			return false;
+		}
+
+		const sampleText = fs.readFileSync( samplePath, 'utf8');
+		clause.parse(sampleText);
+		const parseResult = clause.getData();
+
+		outputChannel.appendLine( 'template' );
+		outputChannel.appendLine( '========' );
+		outputChannel.appendLine( template.getIdentifier() );
+		outputChannel.appendLine( '' );
+
+		outputChannel.appendLine( 'sample.md parse result' );
+		outputChannel.appendLine( '======================' );
+		outputChannel.appendLine( JSON.stringify(parseResult, null, 2) );
+		outputChannel.appendLine( '' );
+
+		const request = JSON.parse(fs.readFileSync( requestPath, 'utf8'));
+
+		outputChannel.appendLine( 'request.json' );
+		outputChannel.appendLine( '============' );
+		outputChannel.appendLine( JSON.stringify(request, null, 2) );
+		outputChannel.appendLine( '' );
+
+		const state = JSON.parse(fs.readFileSync( requestPath, 'utf8'));
+
+		outputChannel.appendLine( 'state.json' );
+		outputChannel.appendLine( '==========' );
+		outputChannel.appendLine( JSON.stringify(state, null, 2) );
+		outputChannel.appendLine( '' );	
+
+		const engine = new Engine();
+		const result = await engine.trigger(clause, request, state, null );
+
+		outputChannel.appendLine( 'response' );
+		outputChannel.appendLine( '========' );
+		outputChannel.appendLine( JSON.stringify(result, null, 2) );
+		outputChannel.appendLine( '' );
+
 		return true;
 	}
 	catch(error) {